requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
requests
python-jose[cryptography]
passlib[bcrypt]
